% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccur.R
\name{cooccur}
\alias{cooccur}
\title{Calculate Co-occurrence Matrix for ICD Codes}
\usage{
cooccur(
  data,
  id,
  code,
  time,
  window = NA,
  pll_njobs = NULL,
  pll_mc.cores = NULL,
  pll_sbatch_opt = NULL,
  out_dir = getwd(),
  output_file = "final_cooccurrence_matrix.rds"
)
}
\arguments{
\item{data}{A data frame containing patient data, will be set to dat.table for optimization.}

\item{id}{The column name representing patient ID.}

\item{code}{The column name representing ICD codes.}

\item{time}{The column name representing the time variable.}

\item{window}{The time window for considering co-occurrence (optional).}

\item{pll_njobs}{Number of parallel jobs (optional).}

\item{pll_mc.cores}{Number of CPU cores for parallel processing (optional).}

\item{pll_sbatch_opt}{Slurm job options,account and partition of user's hpc (optional).}

\item{out_dir}{Directory to save the output file (default is the current working directory).}

\item{output_file}{Name of the output file (default is "final_cooccurrence_matrix.rds").}
}
\value{
A co-occurrence matrix saved to the specified output file.
}
\description{
Calculate Co-occurrence Matrix for ICD Codes
}
\details{
The function calculates a co-occurrence matrix for ICD codes, based on the time window for each patient's data. It supports parallel processing using the \code{slurmR} package with user provide high performance comouter account and partition.
}
\examples{
data("example_data")
cooccur(data = example_data, id = "id", code = "code", time = "time", window = 30, output_file = "final_cooccurrence_matrix.rds")
}
