% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccur_hpc.R
\name{cooccur_hpc}
\alias{cooccur_hpc}
\title{Calculate Co-occurrence Matrix for medical codes on a high-performance computing (HPC) cluster}
\usage{
cooccur_hpc(
  data,
  id,
  code,
  time,
  window = NA,
  pll_njobs = NULL,
  pll_mc.cores = NULL,
  pll_sbatch_opt = NULL,
  out_dir = getwd(),
  output_file = "final_cooccurrence_matrix.rds"
)
}
\arguments{
\item{data}{A data frame containing patient data, will be set to dat.table for optimization.}

\item{id}{The column name representing patient ID.}

\item{code}{The column name representing ICD codes.}

\item{time}{The column name representing the time variable.}

\item{window}{The time window for considering co-occurrence (optional).}

\item{pll_njobs}{Number of parallel jobs (optional).}

\item{pll_mc.cores}{Number of CPU cores for parallel processing (optional).}

\item{pll_sbatch_opt}{Slurm job options,account and partition of user's hpc (optional).}

\item{out_dir}{Directory to save the output file (default is the current working directory).}

\item{output_file}{Name of the output file (default is "final_cooccurrence_matrix.rds").}
}
\value{
A co-occurrence matrix saved to the specified output file.

#' @details
This function is designed for HPC environments using the \code{slurmR} package for job scheduling.
It splits the data by \code{id}, processes each subset in parallel across multiple Slurm jobs,
and computes co-occurrences for the \code{code} values within the specified \code{window}.
The final matrix is the sum of co-occurrences across all subsets.
}
\description{
This function computes a co-occurrence matrix for a dataset that containing medical codes, based on the time window for each patient's data.
It is optimized for high-performance computing (HPC) environments using the \code{slurmR} package.
}
\details{
The function calculates a co-occurrence matrix for medical codes, based on the time window for each patient's data. It supports parallel processing using the \code{slurmR} package with user provide high performance comouter account and partition.
}
\examples{
#data("example_data")
# cooccur(data = data, id = "id",
# code = "code", time = "time", window = NA,
# pll_njobs = 50,
# pll_mc.cores = 40,
# pll_sbatch_opt = list(account = "owner-guest", partition = "notchpeak-guest"),
# out_dir = "~/results"
# output_file = "final_cooccurrence_matrix.rds")
}
