% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccur_local.R
\name{cooccur_local}
\alias{cooccur_local}
\title{Compute a Co-Occurrence Matrix on your local machine}
\usage{
cooccur_local(data, id_col, time_col, code_col, window = NA)
}
\arguments{
\item{data}{A \code{data.table} or \code{data.frame} containing the dataset.}

\item{id_col}{A character string specifying the column name for the unique identifier (e.g., patient ID).}

\item{time_col}{A character string specifying the column name for the time variable.}

\item{code_col}{A character string specifying the column name for the categorical variable of interest (e.g., event codes).}

\item{window}{An optional numeric value specifying the time window for calculating co-occurrences. If \code{NA}, the entire dataset for each ID is used.}
}
\value{
A square matrix where rows and columns correspond to unique codes in \code{code_col}, and the values represent the co-occurrence counts for the entire dataset.
}
\description{
This function computes a co-occurrence matrix for a dataset that containing medical codes, based on the time window for each patient's data.
It uses parallel processing to speed up computation and can be used in either windows or linux operating systems.
}
\details{
The function splits the data by \code{id_col}, processes each group in parallel, and computes the co-occurrences
for the \code{code_col} values within the specified \code{window} based on the \code{time_col}. The final matrix is the sum
of co-occurrences across all groups.
}
\examples{
data("example_data")

co_matrix <- cooccur_local(
  data = example_data,
  id_col = "id",
  time_col = "time",
  code_col = "code",
  window = 3
)
# print(co_matrix)

}
