% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Simulated ICD-9 Data for example purposes}
\format{
A data frame with 57,518 rows and 3 variables:
\describe{
\item{id}{A unique identifier for each patient.}
\item{code}{ICD-9 codes associated with the patient, in the format xxx.xx.}
\item{time}{The time point for each event for a patient. Each patient may associated with multiple events.}
}
}
\source{
Simulated data generated for testing purposes.
}
\usage{
example_data
}
\description{
A dataset containing ICD-9 codes and associated time data for
10,000 unique patients. This dataset was generated for testing and example
purposes. It contains 10,000 unique patient IDs, 1000 unique ICD-9 codes, and time sequences
for each patient's recorded event.
}
\keyword{datasets}
