% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmi_df.R
\name{pmi_df}
\alias{pmi_df}
\title{Calculate Pointwise Mutual Information (PMI) from Co-occurrence Data}
\usage{
pmi_df(cooccur, singletons, my.smooth = 0.75)
}
\arguments{
\item{cooccur}{A data frame with three columns: \code{code1}, \code{code2}, and \code{joint_count}. \code{code1} and \code{code2} represent the pairs of codes, and \code{joint_count} is the count of their co-occurrence.}

\item{singletons}{A list containing two data frames: \code{marg_word} and \code{marg_context}. Each data frame should have two columns: \code{code} and \code{marg_count}, representing the marginal count of each code in the word and context, respectively. \code{singletons} should also contain a scalar \code{D} for the total sum of counts.}

\item{my.smooth}{A numeric value used for smoothing the context marginal count. Default is \code{0.75}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{code1}: The first code in the pair.
\item \code{code2}: The second code in the pair.
\item \code{joint_count}: The co-occurrence count for the pair.
\item \code{W}: The marginal count for \code{code1}.
\item \code{C}: The marginal count for \code{code2}.
\item \code{PMI}: The Pointwise Mutual Information score for the code pair.
}
}
\description{
This function calculates the Pointwise Mutual Information (PMI) score between pairs of codes based on their co-occurrence count and marginal counts and provide the information in the format of data frame. The PMI is used to measure the association between two codes, adjusting for their individual frequencies in the data.
}
\examples{
# Example usage
pmi_results <- pmi_df((cooccur=cooccur_eicu,singletons = sg_eicu,my.smooth = 0.75)

}
