% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppmi_matrix.R
\name{sppmi_matrix}
\alias{sppmi_matrix}
\title{Create a Sparse SPPMI Matrix}
\usage{
sppmi_matrix(pmi, k = 10)
}
\arguments{
\item{pmi}{A data frame containing the PMI values for code pairs. The data frame should have the following columns:
\itemize{
\item \code{code1}: The first code in the pair.
\item \code{code2}: The second code in the pair.
\item \code{PMI}: The Pointwise Mutual Information score for the code pair.
}}

\item{k}{A numeric value for the smoothing constant. Default is 10. The SPPMI is calculated as \code{PMI - log(k)} for each pair of codes.
SPPMI values below 0 are replaced with 0.}
}
\value{
A sparse matrix with the SPPMI values. The matrix has codes as both rows and columns, and the values represent the SPPMI between pairs of codes.
The matrix is symmetric, i.e., \code{sppmi[i,j]} is equal to \code{sppmi[j,i]}.
}
\description{
This function calculates the Shifted Positive Pointwise Mutual Information (SPPMI) for each pair of codes and returns a sparse matrix representing the SPPMI values. The SPPMI is calculated by subtracting the logarithm of a smoothing constant \code{k} from the PMI values, and then applying a threshold to ensure non-negative values. The resulting matrix is symmetric, with codes as both rows and columns.
}
\examples{
# Example
pmi_data <- data.frame(
  code1 = c("A", "B", "A", "C"),
  code2 = c("B", "C", "C", "A"),
  PMI = c(0.5, 1.2, 0.7, 0.9)
)

# Compute SPPMI matrix
sppmi_matrix_result <- sppmi_matrix(pmi_data, k = 10)

# View the resulting sparse matrix
# sppmi_matrix_result
#' @import Matrix
}
